split more out of control.ts
